% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_except.R
\name{remove_except}
\alias{remove_except}
\title{Remove all files from environment except for the files listed as an argument}
\usage{
remove_except(files)
}
\arguments{
\item{files}{List or vector of filenames that should not be removed. Please note that filenames should be quoted}
}
\value{
removes files from global environment
}
\description{
This function will clear the environment with the exception of the files listed.
Please note that the current version of this function will remove datasets, vectors and functions.
The files that should remain in memory can be listed as a vector or as a list.
}
\examples{
\dontrun{
iris_data <- iris
x <- 1
y <- 2
remove_except("iris_data")
remove_except(c("iris_data", "x")
remove_except(list("iris_data", "x")
}


}
