% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_except.R
\name{rm_except}
\alias{rm_except}
\title{Remove all files from environment except for the files listed as an argument}
\usage{
rm_except(objects, remove_action = TRUE)
}
\arguments{
\item{objects}{List or vector of object names that should not be removed. Please note that filenames should be quoted}

\item{remove_action}{By default, remove_action is TRUE and will remove all files except the ones listed in objects.
If FALSE, the function will return a list of all functions that will be removed.}
}
\value{
removes files from global environment
}
\description{
This function will clear the environment with the exception of the files listed.
Please note that the current version of this function will remove datasets, vectors and functions.
The files that should remain in memory can be listed as a vector or as a list.
}
\examples{
\dontrun{
iris_data <- iris
x <- 1
y <- 2
remove_except("iris_data")
remove_except(c("iris_data", "x")
remove_except(list("iris_data", "x")
}


}
